Cambio en Entity (dominio) que realizaba llamadas a a la api. Ahora esta funcionalidad se encuentra en External (data)
