+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/MovieRepositoryImpl.kt
@@ -4,7 +4,7 @@ import edu.dyds.movies.data.external.TheMoviesDataBase
 import edu.dyds.movies.data.local.LocalCache
 import edu.dyds.movies.domain.entity.Movie
 import edu.dyds.movies.domain.entity.QualifiedMovie
-import edu.dyds.movies.domain.entity.RemoteMovie
+import edu.dyds.movies.data.external.RemoteMovie
 import edu.dyds.movies.domain.repository.MovieRepository
 
 class MovieRepositoryImpl( private val cache: LocalCache, private val theMoviesDataBase : TheMoviesDataBase) : MovieRepository {
