+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/domain/repository/MovieRepository.kt
@@ -2,7 +2,7 @@ package edu.dyds.movies.domain.repository
 
 import edu.dyds.movies.domain.entity.Movie
 import edu.dyds.movies.domain.entity.QualifiedMovie
-import edu.dyds.movies.domain.entity.RemoteMovie
+import edu.dyds.movies.data.external.RemoteMovie
 
 interface MovieRepository {
     suspend fun getPopularMovies() : List<RemoteMovie>
