+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/domain/usecase/GetPopularMoviesUseCase.kt
@@ -3,7 +3,7 @@ package edu.dyds.movies.domain.usecase
 import edu.dyds.movies.domain.repository.MovieRepository
 import edu.dyds.movies.domain.entity.Movie
 import edu.dyds.movies.domain.entity.QualifiedMovie
-import edu.dyds.movies.domain.entity.RemoteMovie
+import edu.dyds.movies.data.external.RemoteMovie
 
 private const val MIN_VOTE_AVERAGE = 6.0
